% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advr38pkg-package.R
\docType{package}
\name{advr38pkg-package}
\alias{advr38pkg}
\alias{advr38pkg-package}
\title{advr38pkg: Advanced R Course}
\description{
Associated R package with the materials for the Advanced R course at https://privefl.github.io/advr38book/.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/privefl/advr38pkg/issues}
}

}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
