% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_walk_neg_prop}
\alias{random_walk_neg_prop}
\title{Random Walk}
\usage{
random_walk_neg_prop(N)
}
\arguments{
\item{N}{Size of the walk.}
}
\description{
Negative proportion of a Gaussian random walk.
}
\examples{
set.seet(1)
random_walk_neg_prop(1e6)

}
