% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprog.R
\name{split_ind}
\alias{split_ind}
\title{Split indices in K groups}
\usage{
split_ind(ind, K)
}
\arguments{
\item{ind}{Vector of indices to split.}

\item{K}{Number of groups to make.}
}
\value{
A list of \code{K} groups of indices.
}
\description{
Split indices in K groups
}
\examples{
advr38pkg::split_ind(1:12, 5)
advr38pkg::split_ind(1:40, 3)
advr38pkg::split_ind(10, 5)
}
