% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprog.R
\name{sum_every}
\alias{sum_every}
\title{Get the sums of blocks}
\usage{
sum_every(x, n)
}
\arguments{
\item{x}{Numeric vector.}

\item{n}{Positive number as the size of the blocks.}
}
\value{
The sums of each block.
}
\description{
Get the sums of blocks of \code{n} elements of a vector \code{x}.
}
\examples{
sum_every(1:10, 2)
sum_every(1:10, 5)
}
